Allocate a new buffer per request packet
Request Packet types are passed directly to ServeDHCP().  If a
ServeDHCP() implementation saves a reference to the Packet or return
values to Packet.*Addr(), which are all just slices of the same Packet
buffer, they become stale as soon as the next request Packet is read
over the top of the same buffer

By allocating a new request Packet every time, the references will
remain consistent forever at the cost of some garbage collection.